pyproject.toml
[tool.poetry]
name = "pyready"
version = "0.1.0"
description = "Python project environment checker and governance tool with deterministic analysis"
authors = ["Aditya Mallela <your.email@example.com>"]
readme = "README.md"
license = "MIT"
homepage = "https://github.com/yourusername/pyready"
repository = "https://github.com/yourusername/pyready"
keywords = ["cli", "python", "environment-check", "governance", "developer-tools", "project-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Quality Assurance",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
packages = [{include = "pyready"}]

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^2.5.0"
tomli = "^2.0.1"
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.7.0"
httpx = {version = "^0.27.0", optional = true}
fastapi = {version = "^0.109.0", optional = true}
uvicorn = {version = "^0.27.0", optional = true}
pyyaml = "^6.0.3"
python-dotenv = "^1.2.1"
groq = {version = "^0.4.0", optional = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"

[tool.poetry.extras]
api = ["fastapi", "uvicorn", "httpx"]
ai = ["groq"]
all = ["fastapi", "uvicorn", "httpx", "groq"]

[tool.poetry.scripts]
pyready = "pyready.cli.main:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
